When AnkhSVN is not (yet) enabled, place the Pending Changes window in a generic disabled state which allows easy access to the SCC enable function like that in Team Foundation server.
If the ToolWindow state stored in the Visual Studio settings indicates that
the ToolWindow should be opened when Visual Studio starts, we now only load
in this not-active state and switch to active when the SCC provider is really
activated (which assumes a working UI shell).

This should resolve issues in 3th party and Microsoft extensions that assume
language services can only be instantiated when the UI shell is available.

* src/Ankh.UI/Ankh.UI.csproj
  Add new files

* src/Ankh.UI/PendingChanges/PendingActivationPage.cs
* src/Ankh.UI/PendingChanges/PendingActivationPage.Designer.cs
* src/Ankh.UI/PendingChanges/PendingActivationPage.resx
  New pending activation window; showing an enable SCC provider link.

* src/Ankh.UI/PendingChanges/PendingChangesPage.cs
  Make a few properties invisible for the designer.

* src/Ankh.UI/PendingChanges/PendingChangesToolControl.cs
  Check if we should activate our support, or just show a stub page instead.